avviddec: Remap buffer with only READ flag before pushing
No need to keep WRITE access when the decoder has reconstructed the
frame. Not mapping the memory as WRITE reduces the number of situations
where memcpy is needed. E.g gst_buffer_make_writable may not get a
exclusive reference to the memory if mapped as WRITE which leads to
memcpy.